# contracts/http/openapi.yaml
openapi: 3.1.2
info:
  title: Labbit HTTP API
  version: 0.1.5
  description: |
    Labbit Browser ↔ SaaS HTTP 계약의 기계 판독 가능한 SSOT입니다.

    이 계약은 Class, LabSpec, LabExecution, LabInstance, Operation처럼
    클라이언트가 알아야 하는 안정적인 제품 리소스만 노출합니다.
    PostgreSQL polling, SQS/Kafka, Valkey, Istio, Kubernetes workload 경계,
    향후 서비스 분리와 같은 내부 실행 기술은 HTTP 계약에 노출하지 않습니다.

    호환성 원칙:
    - 호환되지 않는 HTTP 변경은 새로운 Major API 경로에서 제공합니다.
    - v1 안에서는 기존 클라이언트를 깨지 않는 필드를 추가할 수 있으며,
      클라이언트는 사용하지 않는 응답 필드를 무시해야 합니다.
    - 리소스 ID는 내부 생성 방식을 가정하지 않는 opaque string으로 취급합니다.
    - 오래 걸리는 Provider 변경 작업은 Operation 리소스로 추적합니다.
    - 오류 code, Operation stage, lifecycle status는 향후 값이 추가될 수 있습니다.
servers:
  - url: /api/v1
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: 사전 생성 Local Account로 로그인
      operationId: login
      security: []
      requestBody:
        required: true
        description: 운영자가 사전에 생성한 Local Account의 로그인 정보
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '204':
          description: 인증에 성공했으며 기존 값을 재사용하지 않는 새 브라우저 세션 Cookie를 발급했습니다.
          headers:
            Set-Cookie:
              $ref: '#/components/headers/SessionSetCookie'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CsrfRejected'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/logout:
    post:
      tags: [Auth]
      summary: 현재 로그인 세션 종료
      operationId: logout
      responses:
        '204':
          description: 현재 세션을 종료하고 세션 Cookie를 제거했습니다.
          headers:
            Set-Cookie:
              $ref: '#/components/headers/SessionClearCookie'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CsrfRejected'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /me:
    get:
      tags: [Auth]
      summary: 현재 로그인 사용자 조회
      operationId: getMe
      responses:
        '200':
          description: 현재 사용자와 소속 Organization의 안전한 인증 컨텍스트입니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /classes:
    get:
      tags: [Classes]
      summary: 접근 가능한 Class 목록 조회
      operationId: listClasses
      responses:
        '200':
          description: 현재 사용자가 접근할 수 있는 Class 목록입니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /classes/{classId}:
    get:
      tags: [Classes]
      summary: Class 상세와 현재 사용자 컨텍스트 조회
      operationId: getClass
      parameters:
        - $ref: '#/components/parameters/ClassId'
      responses:
        '200':
          description: Class 상세, 현재 사용자의 역할, 활성 실습 실행과 자신의 실습 환경 요약입니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: |
            요청한 classId에 해당하는 Class가 존재하지 않습니다.
            서버가 발급하지 않은 형식의 classId도 같은 응답입니다.
            Class가 존재하지만 현재 사용자가 접근할 수 없으면 이 응답이 아니라 403입니다.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /lab-instances/{labInstanceId}/terminal-targets:
    get:
      tags: [TerminalSessions]
      summary: 본인 LabInstance에서 Terminal을 열 수 있는 VM 목록 조회
      operationId: listTerminalTargets
      description: |
        현재 사용자가 소유한 LabInstance에서 TerminalSession을 열 수 있는 VM의 목록입니다.
        `POST /lab-instances/{labInstanceId}/terminal-sessions`의 `targetVmKey`는 이 응답의 `items[].vmKey`를
        그대로 사용합니다. 클라이언트가 `"workspace"` 같은 key를 하드코딩하거나 `role`/`instanceIndex`로 만들어서는 안 됩니다.

        Terminal은 Editor/Preview처럼 Workspace VM 하나에 한정되지 않고 LabInstance 안의 여러 VM 중에서 선택할 수 있으므로
        `items`는 Workspace VM을 포함한 모든 target입니다. `workspaceVmKey`는 `items` 중 Workspace VM을 가리키는
        기본 선택 hint입니다.

        목록의 원본은 해당 LabExecution에 고정된 immutable resolved CreationSnapshot입니다. 최신 LabSpec을 다시 해석하지 않으며
        Provider 내부 topology(Provider Server ID, Connector ID, private IP, 이미지·flavor 등)는 응답에 포함하지 않습니다.
        이 목록은 논리적으로 선택할 수 있는 VM이며 VM이 지금 사용 가능하다는 보증이 아닙니다. 실제 TerminalSession 생성 가능 여부는
        생성 시점의 LabInstance status, 현재 generation, ProviderResource로 다시 판단합니다.

        TerminalSession 생성과 같은 권한 경계를 사용합니다. 하나라도 맞지 않으면 거절합니다.

        - LabInstance가 현재 사용자의 것이며 같은 Organization에 속합니다(`403`). 강사나 Organization `ADMIN`도
          다른 사용자의 LabInstance는 조회할 수 없습니다.
        - 사용자가 LabInstance가 속한 Class의 현재 ClassMembership(INSTRUCTOR 또는 STUDENT)을 가집니다(`403`).
        - LabInstance status가 `READY`입니다(`409`, code `lab_instance_not_ready`).

        저장된 CreationSnapshot을 해석할 수 없거나 서로 모순되면(빈 key, 중복 `vmKey`, `workspaceVmKey`가 `items`에 없음 등)
        일부만 반환하지 않고 `500`으로 거절합니다. 응답은 `Cache-Control: no-store`입니다.
      parameters:
        - $ref: '#/components/parameters/LabInstanceId'
      responses:
        '200':
          description: Terminal을 열 수 있는 VM 목록입니다.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalTargetList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /lab-instances/{labInstanceId}/terminal-sessions:
    post:
      tags: [TerminalSessions]
      summary: 본인 LabInstance의 VM에 TerminalSession 생성
      operationId: createTerminalSession
      description: |
        현재 사용자가 소유한 LabInstance의 현재 generation에 있는 VM 하나에 새 TerminalSession을 만들고,
        Browser Terminal WSS(`/realtime/v1/terminal`)에 attach할 때 사용하는 terminal-scoped opaque
        `sessionToken`을 발급합니다.

        서버는 요청의 `targetVmKey`만 받아 현재 DB 상태에서 대상 VM과 Connector를 결정합니다.
        `targetVmKey`는 `GET /lab-instances/{labInstanceId}/terminal-targets`가 돌려준 `vmKey`여야 하며,
        서버는 조회 결과를 신뢰하지 않고 생성 시점에 다시 판단합니다.
        Provider Server ID, Connector ID, generation은 요청으로 받지 않으며 신뢰하지 않습니다.
        다음을 모두 만족해야 TerminalSession을 만들 수 있고, 하나라도 맞지 않으면 어떤 side effect도 만들기 전에 거절합니다.

        - LabInstance가 현재 사용자의 것이며 같은 Organization에 속합니다(`403`).
        - 사용자가 LabInstance가 속한 Class의 현재 ClassMembership(INSTRUCTOR 또는 STUDENT)을 가집니다(`403`).
        - LabInstance status가 `READY`입니다(`409`, code `lab_instance_not_ready`).
        - `targetVmKey`가 해당 LabExecution의 immutable CreationSnapshot에 있는 VM입니다(`422`, code `invalid_terminal_target`).
          현재 generation에 같은 이름의 ProviderResource가 있어도 CreationSnapshot에 없는 key는 target이 아닙니다.
        - 그 VM에 해당하는 현재 generation의 SERVER ProviderResource가 `PRESENT`입니다.
          리소스가 없거나 `PRESENT`가 아니면 `409`, code `terminal_target_unavailable`입니다.

        강사가 학생 LabInstance의 Terminal을 직접 여는 기능은 이 계약 범위가 아닙니다.

        TerminalSession 생성은 동기적으로 완료됩니다. 서버가 Connector에 PTY 준비를 요청하고, Connector의 Terminal Data
        채널이 실제로 연결된 것을 확인한 뒤에만 `201`을 반환합니다. Connector를 사용할 수 없거나 준비에 실패하거나
        시간 안에 끝나지 않으면 TerminalSession은 성공 상태로 남지 않고 `503`으로 실패합니다
        (code `connector_unavailable` 또는 `terminal_open_failed`).

        이 배포 구성이 Terminal Relay를 제공하지 않으면 `503`(code `terminal_unavailable`)입니다.

        응답에는 Browser가 연결하는 데 필요한 최소 정보만 포함합니다. WSS endpoint와 subprotocol은 고정이며
        세션마다 URL을 돌려주지 않습니다. 응답은 `Cache-Control: no-store`입니다.
      parameters:
        - $ref: '#/components/parameters/LabInstanceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTerminalSessionRequest'
      responses:
        '201':
          description: TerminalSession이 만들어졌고 Browser가 attach할 수 있습니다.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalSession'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /terminal-sessions/{terminalSessionId}:
    delete:
      tags: [TerminalSessions]
      summary: TerminalSession 명시적 종료
      operationId: closeTerminalSession
      description: |
        사용자가 자신의 TerminalSession을 명시적으로 종료합니다. Browser WSS의 일시적 단절과 달리
        60초 reconnect grace를 시작하지 않고 즉시 `ENDED`로 전환하며, 이 TerminalSession의 attach token은
        더 이상 사용할 수 없습니다. 열려 있는 Browser WSS에는 `TERMINAL_SESSION_ENDED`(reason `SESSION_CLOSED`)를 전달합니다.

        이미 종료된 TerminalSession에 대한 반복 요청도 `204`입니다(멱등).

        서버는 Connector에 PTY 종료(`TERMINAL_CLOSE`)를 요청합니다. Connector를 사용할 수 없어 전달하지 못해도
        Labbit의 TerminalSession은 `ENDED`이며 이후 attach할 수 없습니다. 남은 PTY 정리는 Connector lifecycle의 책임입니다.
      parameters:
        - $ref: '#/components/parameters/TerminalSessionId'
      responses:
        '204':
          description: TerminalSession이 종료되었습니다(이미 종료된 경우 포함).
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /lab-instances/{labInstanceId}/files/tree:
    get:
      tags: [WorkspaceFiles]
      summary: 본인 Workspace의 디렉터리 하나 나열
      operationId: listWorkspaceFiles
      description: |
        현재 사용자가 소유한 LabInstance의 Workspace VM에서 디렉터리 하나의 직계 항목을 나열합니다.
        재귀 조회는 제공하지 않으며 클라이언트가 디렉터리를 펼칠 때마다 그 디렉터리의 `path`로 다시 조회합니다.

        Browser는 VM, Connector, Provider Server, generation, private address를 선택하거나 알 필요가 없습니다. 서버가 LabExecution의
        immutable CreationSnapshot의 `workspaceVmKey`와 현재 generation의 `PRESENT` `SERVER` ProviderResource로 대상 VM을
        결정하고 Connector와 Workspace VM의 SFTP를 거쳐 조회합니다. Browser가 VM private network나 SFTP에 직접 접근하지 않습니다.

        권한 경계는 TerminalSession 생성과 같습니다. 하나라도 맞지 않으면 Connector를 호출하기 전에 거절합니다.

        - LabInstance가 현재 사용자의 것이며 같은 Organization에 속합니다(`403`). 강사나 Organization `ADMIN`도
          다른 사용자의 LabInstance는 조회할 수 없습니다.
        - 사용자가 LabInstance가 속한 Class의 현재 ClassMembership(INSTRUCTOR 또는 STUDENT)을 가집니다(`403`).
        - LabInstance status가 `READY`입니다(`409`, code `lab_instance_not_ready`).
        - 현재 generation에 Workspace VM의 `PRESENT` `SERVER` ProviderResource가 정확히 하나 있습니다. 없으면 `409`,
          code `workspace_target_unavailable`입니다. 저장된 CreationSnapshot이나 ProviderResource 관계가 모순이면 `500`입니다.

        `path`는 `path` query parameter에 정의한 경로 규칙을 따릅니다. 생략하거나 빈 값이면 Workspace root입니다.
        `items`는 `name`의 UTF-8 byte 순서 오름차순입니다. 항목의 `kind`는 `file`과 `directory`뿐이며 그 밖의 종류의
        항목(예: symlink, device)이나 위 경로 규칙으로 표현할 수 없는 이름은 목록에 포함하지 않습니다. uid/gid, POSIX mode,
        크기, 수정 시각 같은 파일 metadata는 노출하지 않습니다.

        지정한 경로가 없으면 `404`(code `path_not_found`), 디렉터리가 아니면 `422`(code `path_not_directory`)입니다.
        목록이 Connector 전송 한도를 넘으면 `413`(code `directory_too_large`)입니다. Connector나 VM을 사용할 수 없으면
        `503`입니다(code `connector_unavailable` 또는 `file_transport_unavailable`).

        응답에는 파일 본문이 없지만 Workspace 구조는 학생 작업물이므로 `Cache-Control: no-store`입니다.
      parameters:
        - $ref: '#/components/parameters/LabInstanceId'
        - $ref: '#/components/parameters/WorkspaceDirectoryPath'
      responses:
        '200':
          description: 요청한 디렉터리의 직계 항목입니다.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceFileTree'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /lab-instances/{labInstanceId}/files/content:
    get:
      tags: [WorkspaceFiles]
      summary: 본인 Workspace의 UTF-8 text 파일 읽기
      operationId: readWorkspaceFile
      description: |
        현재 사용자가 소유한 LabInstance의 Workspace VM에서 UTF-8 text 파일 하나를 읽습니다.
        대상 VM 결정과 권한 경계는 `GET /lab-instances/{labInstanceId}/files/tree`와 같습니다.

        응답의 `ETag`는 이 읽기 시점의 파일 revision을 나타내는 opaque 값이며 `PUT`의 `If-Match`로 되돌려 보냅니다.
        revision을 만드는 방식과 형식은 호환성 계약이 아니므로 클라이언트는 해석하지 않습니다.

        MVP는 UTF-8 text editor만 지원합니다. 파일 본문이 유효한 UTF-8이 아니면 `422`(code `unsupported_encoding`),
        NUL byte를 포함한 binary이면 `422`(code `binary_content`)이며 본문을 응답에 싣지 않습니다. 파일이 설정된 크기
        한도를 넘으면 `413`(code `file_too_large`)입니다. 한도의 수치는 계약이 아니며 배포 구성에 따라 달라질 수 있습니다.
        지정한 경로가 없으면 `404`(code `path_not_found`), 일반 파일이 아니면 `422`(code `path_not_file`)입니다.
        Workspace VM 파일 시스템이 접근을 거절하면 `403`(code `file_permission_denied`)이며 이는 Labbit 권한과 구분되는 VM 내부 권한입니다.

        파일 본문과 경로는 PostgreSQL과 log, trace, metric에 저장하거나 기록하지 않습니다. 응답은 `Cache-Control: no-store`입니다.
      parameters:
        - $ref: '#/components/parameters/LabInstanceId'
        - $ref: '#/components/parameters/WorkspaceFilePath'
      responses:
        '200':
          description: 파일 본문과 revision입니다.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceFileContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

    put:
      tags: [WorkspaceFiles]
      summary: 본인 Workspace의 기존 UTF-8 text 파일 저장
      operationId: saveWorkspaceFile
      description: |
        기존 파일 하나의 본문을 교체합니다. 새 파일 생성, 이름 변경, 삭제는 이 계약 범위가 아닙니다. 지정한 경로에 파일이 없으면
        `404`(code `path_not_found`)이며 파일을 만들지 않습니다.

        `If-Match`에 같은 파일을 읽을 때 받은 `ETag` 하나를 반드시 보내야 합니다. 서버는 저장 직전에 현재 파일 revision과
        비교해 일치할 때만 저장하고, 다르면 덮어쓰지 않고 `412`(code `stale_revision`)로 거절합니다. 클라이언트는 다시 읽어
        변경 내용을 합친 뒤 저장합니다. `If-Match`가 없으면 `400`(code `if_match_required`), 하나의 strong entity-tag가
        아니면(`*`, 목록, weak validator 포함) `400`(code `invalid_if_match`)입니다.

        대상 VM 결정과 권한 경계는 `GET /lab-instances/{labInstanceId}/files/tree`와 같고, unsafe method이므로 Origin 검증도 적용합니다.

        요청 본문은 `{"content": "..."}`입니다. 유효한 UTF-8이 아니거나 짝이 맞지 않는 surrogate escape가 있으면
        `422`(code `unsupported_encoding`), NUL을 포함하면 `422`(code `binary_content`), 설정된 크기 한도를 넘으면
        `413`(code `file_too_large`)이며 Connector를 호출하기 전에 거절합니다.

        성공하면 새 revision을 `ETag`로 돌려줍니다. Connector 응답을 받지 못해 저장 여부를 알 수 없으면(전송 이후 연결 단절,
        시간 초과) 서버는 자동으로 다시 저장하지 않고 `503`(code `file_save_outcome_unknown`)으로 알립니다. 클라이언트는
        파일을 다시 읽어 저장 여부를 확인합니다. 저장 중 Workspace VM의 generation이 바뀌었거나 LabInstance가 더 이상 `READY`가
        아니면 성공으로 처리하지 않고 `409`(code `workspace_target_changed`)입니다.

        파일 본문과 경로는 PostgreSQL과 log, trace, metric에 저장하거나 기록하지 않습니다.
      parameters:
        - $ref: '#/components/parameters/LabInstanceId'
        - $ref: '#/components/parameters/WorkspaceFilePath'
        - $ref: '#/components/parameters/WorkspaceFileIfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveWorkspaceFileRequest'
      responses:
        '200':
          description: 파일을 저장했습니다. `ETag`는 저장된 파일의 새 revision입니다.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceFileSaved'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /preview-sessions/{previewSessionId}:
    delete:
      tags: [PreviewSessions]
      summary: PreviewSession 명시적 종료
      operationId: closePreviewSession
      description: |
        사용자가 자신의 PreviewSession을 명시적으로 종료합니다. 즉시 Preview Origin의 Cookie 인증이 무효가 되고(이후 요청은 `401`)
        열려 있는 Workspace tunnel이 닫힙니다.

        이미 종료되었거나 만료된 PreviewSession에 대한 반복 요청도 `204`입니다(멱등). 서버가 기억하지 못하는 ID(알 수 없는 ID,
        서버 재시작 뒤)는 `404`, 다른 사용자의 PreviewSession은 `403`입니다.

        서버는 Connector에 TCP forwarding 종료(`PREVIEW_CLOSE`)를 요청합니다. Connector를 사용할 수 없어 전달하지 못해도
        Labbit의 PreviewSession은 종료된 상태이며 이후 사용할 수 없습니다.
      parameters:
        - $ref: '#/components/parameters/PreviewSessionId'
      responses:
        '204':
          description: PreviewSession이 종료되었습니다(이미 종료된 경우 포함).
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-labbit-session
      description: |
        Labbit Backend가 발급하는 opaque 브라우저 세션 식별자입니다.

        Production 및 production-like HTTPS 환경의 Cookie 속성은
        Secure, HttpOnly, SameSite=Lax, Path=/를 사용하고 Domain은 지정하지 않습니다.
        v0.1에서는 Max-Age/Expires를 두지 않는 non-persistent browser Cookie로 사용합니다.
        브라우저는 Cookie 값을 직접 읽거나 인증 상태의 원본으로 사용하지 않습니다.

        로그인 성공마다 cryptographically-random 새 Session token을 발급하며
        기존 Cookie 값을 인증 후 세션 ID로 재사용하거나 고정하지 않습니다.
        서버 Session의 v0.1 absolute lifetime은 생성 시점부터 8시간이며
        활동에 따른 sliding extension이나 주기적 token renewal은 적용하지 않습니다.
        만료·폐기·비활성 사용자 또는 Password 변경 이전 Session은 인증에 사용할 수 없고
        보호 API는 401을 반환합니다.

        세션 저장소는 HTTP 계약의 구현 상세입니다.
        PostgreSQL에서 다른 저장소로 변경하더라도 Cookie와 인증 의미는 유지합니다.

        POST/PUT/PATCH/DELETE 같은 unsafe method는 SameSite Cookie만 신뢰하지 않고,
        서버가 trusted public origin과 Origin을 정확히 비교합니다.
        Origin이 없는 경우 Referer의 origin을 fallback으로 확인하며
        둘 다 없거나 일치하지 않으면 403으로 거절합니다.
        구체 구현 기준은 docs/backend/auth-session.md를 따릅니다.
    previewCookie:
      type: apiKey
      in: cookie
      name: __Host-labbit-preview
      description: |
        **Preview Origin(SaaS 본 서비스 Origin과 다른 host)에서만** 쓰는 Preview 전용 Cookie입니다. 로그인 Session Cookie와 별개이며
        SaaS 본 서비스 Origin의 인증에 쓰이지 않습니다. 이 API의 `/api/v1` 경로에는 적용되지 않습니다.

        `POST /__labbit/exchange`가 일회용 bootstrap credential을 교환할 때 발급하며 값은 CSPRNG가 만든 32 bytes의 padding 없는
        Base64URL 문자열입니다. 서버는 원문을 저장하지 않고 SHA-256 digest만 보관하고 constant time으로 비교합니다.

        속성은 HttpOnly, Path=/, SameSite=Lax이며 **Domain을 지정하지 않는** host-only Cookie입니다. HTTPS Preview Origin(production 필수)에서는
        Secure이고 이름이 `__Host-labbit-preview`입니다. 개발용 HTTP Preview Origin에서만 `__Host-` prefix를 쓸 수 없으므로
...
  headers:
    SessionSetCookie:
      description: '새 opaque token으로 Secure; HttpOnly; SameSite=Lax; Path=/, Domain 미지정의 non-persistent __Host-labbit-session Cookie를 발급합니다.'
      schema:
        type: string
...
    ETag:
      description: 조건부 수정에 사용하는 opaque 엔터티 버전 값입니다.
      schema:
        type: string
...
  responses:
    BadRequest:
      description: |
        요청 JSON의 문법이나 기본 형식이 올바르지 않습니다.

        Workspace file API에서는 code `invalid_request`(query/body 형식 오류, 필수 `path` 누락, `path` 중복), `invalid_path`(경로 규칙 위반),
        `if_match_required`, `invalid_if_match`입니다.

        PreviewSession 생성과 bootstrap 교환에서는 code `invalid_request`(body 형식 오류, `targetPort`가 1~65535의 정수가 아님 등)입니다.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
...
    LoginRequest:
      type: object
      description: 사전 생성 Local Account 로그인 요청입니다.
      required: [username, password]
      additionalProperties: false
      properties:
        username:
          type: string
          minLength: 1
          description: 운영자가 사전에 생성한 Local Account의 username입니다.
        password:
          type: string
          minLength: 1
          writeOnly: true
          description: 로그인 검증에만 사용하는 password 원문입니다. 저장하거나 로그에 기록하지 않습니다.
...
security:
  - sessionCookie: []